[#139] refactor: 폴더 이름 변경(카멜케이스->케밥케이스)
