+++ src/app/debate-list/[debateId]/components/DebateDetail.tsx
