+++ src/app/debate-list/[debateId]/components/DebateSlider.tsx
