+++ src/app/debate-list/[debateId]/page.server.tsx
