+++ src/app/debate-list/[debateId]/page.tsx
