+++ src/app/debate-list/components/CommentCard.tsx
