+++ src/app/debate-list/components/DebateListCard.tsx
