+++ src/app/debate-list/components/DebatePrevNextNav.tsx
